refactor search controller
